refactor(auth): 将认证从 ScuApiService 拆出为独立的 ScuAuthService
- 新建 lib/services/scu_auth/scu_auth_service.dart
  - 仅负责认证：login / fetchCaptcha / bindSession / logout
  - 持有 _accessToken / _cachedClient / _bindSessionFuture 等认证状态
  - 无 AuthManager 依赖，可独立单测
  - 定义 ScuLoginException / CaptchaResult / kZhjwBase

- 重构 lib/services/scu_api/scu_api_service.dart
  - 只保留 bindAuthManager() / request() / _checkSessionExpiry()
  - 删除 login / fetchCaptcha / bindSession / logout / 认证状态
  - 删除 _warmupJwtSso / _warmupCasSso / _doBindSession / 静态配置
  - 删除 scu_api_models.dart（内容已迁至 scu_auth）
  - 添加 ScuAuthService.requestHeaders 公开 getter 供 extension 使用

- 更新 ScuAuthSession
  - 持有 ScuAuthService 而非 ScuApiService
  - 移除 service getter
  - 新增 authService getter 供外部访问 fetchCaptcha 等

- 更新 AuthManager
  - 新增 apiService 字段（持有 ScuApiService 实例）
  - 构造函数中调 apiService.bindAuthManager(this)
  - 移除 scu.service.bindAuthManager(this)

- 更新 ScuAuthProvider.service getter 返回 _authManager.apiService
- 新增 ScuAuthProvider.authService getter 返回 _authManager.scu.authService

- 更新 scu_login_page.dart 使用 authService.fetchCaptcha()
- 更新 scu_auth_provider.dart autoLogin 使用 _authManager.scu.authService

- 去除循环依赖的延迟绑定 hack (bindAuthManager on service)
- barrel export scu_api_service.dart 新增 scu_auth_service.dart 导出
