+++ lib/services/scu_api/scu_api_classroom.dart
@@ -13,7 +13,7 @@ extension ScuApiClassroom on ScuApiService {
         headers: {
           'Accept': 'text/html,*/*',
           'Referer': '$kZhjwBase/',
-          'User-Agent': ScuApiService._headers['User-Agent']!,
+          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
         },
       );
       final body = resp.body.trim();
@@ -59,7 +59,7 @@ extension ScuApiClassroom on ScuApiService {
         headers: {
           'Accept': 'text/html,*/*',
           'Referer': '$kZhjwBase/',
-          'User-Agent': ScuApiService._headers['User-Agent']!,
+          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
         },
       );
       final body = resp.body.trim();
@@ -94,7 +94,7 @@ extension ScuApiClassroom on ScuApiService {
           'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
           'Referer':
               '$kZhjwBase/student/teachingResources/classroomUseStatus/index',
-          'User-Agent': ScuApiService._headers['User-Agent']!,
+          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
           'X-Requested-With': 'XMLHttpRequest',
         },
         body:
